+++ db/dbformat.cc
@@ -17,6 +17,7 @@
 #include "monitoring/perf_context_imp.h"
 #include "port/port.h"
 #include "util/coding.h"
+#include "util/string_util.h"
 
 namespace rocksdb {
