+++ options/options_helper.h
@@ -39,44 +39,6 @@ static std::map<CompactionPri, std::string> compaction_pri_to_string = {
 
 #ifndef ROCKSDB_LITE
 
-// Returns true if the input char "c" is considered as a special character
-// that will be escaped when EscapeOptionString() is called.
-//
-// @param c the input char
-// @return true if the input char "c" is considered as a special character.
-// @see EscapeOptionString
-bool isSpecialChar(const char c);
-
-// If the input char is an escaped char, it will return the its
-// associated raw-char.  Otherwise, the function will simply return
-// the original input char.
-char UnescapeChar(const char c);
-
-// If the input char is a control char, it will return the its
-// associated escaped char.  Otherwise, the function will simply return
-// the original input char.
-char EscapeChar(const char c);
-
-// Converts a raw string to an escaped string.  Escaped-characters are
-// defined via the isSpecialChar() function.  When a char in the input
-// string "raw_string" is classified as a special characters, then it
-// will be prefixed by '\' in the output.
-//
-// It's inverse function is UnescapeOptionString().
-// @param raw_string the input string
-// @return the '\' escaped string of the input "raw_string"
-// @see isSpecialChar, UnescapeOptionString
-std::string EscapeOptionString(const std::string& raw_string);
-
-// The inverse function of EscapeOptionString.  It converts
-// an '\' escaped string back to a raw string.
-//
-// @param escaped_string the input '\' escaped string
-// @return the raw string of the input "escaped_string"
-std::string UnescapeOptionString(const std::string& escaped_string);
-
-uint64_t ParseUint64(const std::string& value);
-
 Status GetMutableOptionsFromStrings(
     const MutableCFOptions& base_options,
     const std::unordered_map<std::string, std::string>& options_map,
@@ -754,7 +716,6 @@ static std::unordered_map<std::string, InfoLogLevel> info_log_level_string_map =
      {"FATAL_LEVEL", InfoLogLevel::FATAL_LEVEL},
      {"HEADER_LEVEL", InfoLogLevel::HEADER_LEVEL}};
 
-extern const std::string kNullptrString;
 #endif  // !ROCKSDB_LITE
 
 }  // namespace rocksdb
