+++ tools/reduce_levels_test.cc
@@ -11,7 +11,7 @@
 #include "rocksdb/db.h"
 #include "rocksdb/utilities/ldb_cmd.h"
 #include "tools/ldb_cmd_impl.h"
-#include "util/logging.h"
+#include "util/string_util.h"
 #include "util/testharness.h"
 #include "util/testutil.h"
