+++ utilities/date_tiered/date_tiered_test.cc
@@ -14,6 +14,7 @@
 #include "rocksdb/compaction_filter.h"
 #include "rocksdb/utilities/date_tiered_db.h"
 #include "util/logging.h"
+#include "util/string_util.h"
 #include "util/testharness.h"
 
 namespace rocksdb {
